example/django_tutorial/requirements.dev.txt
isort
flake8
pylint
pylint-django
neovim
